% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi-sqr-periodogram.R
\name{chi_sq_periodogram}
\alias{chi_sq_periodogram}
\title{Chi Square Periodogram}
\usage{
chi_sq_periodogram(x, period_range = c(hours(16), hours(32)),
  sampling_rate = 1/mins(1), alpha = 0.05, time_resolution = hours(0.1))
}
\arguments{
\item{x}{numeric vector}

\item{period_range}{vector of size 2 defining minimal and maximal range of period to study (in seconds)}

\item{sampling_rate}{the -- implicitly regular -- sampling rate of x (in hertz)}

\item{alpha}{significance level}

\item{time_resolution}{the resolution of periods to scan}
}
\description{
TODO
}
\examples{
x <- behavr::toy_dam_data(duration = days(5))$activity
s <- chi_sq_periodogram(x)
\dontrun{
ggplot2::ggplot(s, ggplot2::aes(period, power)) +
        ggplot2::geom_line() +
        ggplot2::geom_line (ggplot2::aes(y = signif_level),linetype="dashed", colour="grey") +
        ggplot2::scale_x_time()
}
}
\seealso{
\link[xsp:chiSqPeriodogram]{xsp::chiSqPeriodogram} (code derived from)
}
