% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Find peaks in a periodogram}
\usage{
find_peaks(data, n_peaks = 3, alpha = 0.05)
}
\arguments{
\item{data}{\link[behavr:behavr]{behavr::behavr} table represneted a periodogram, as returned by \link{periodogram}}

\item{n_peaks}{maximal numbers of peak to be detected}

\item{alpha}{the significance threshold of p-values over which peaks are ignored}
}
\value{
\link[behavr:behavr]{behavr::behavr} table that is \code{data} with an extra column \code{peak}.
\code{peak} is filled with \code{NA} values except for rows match a peak.
In which case, they have an integer value corresponding to the rank of the peak (e.g. 1 for the first peak).
}
\description{
Locate the peaks in a pregenerated periodogram. Detection is based on \link[pracma:findpeaks]{pracma::findpeaks}.
Only the significant (with threshold `alpha``) peaks are extracted.
}
\examples{
data(dams_sample)
per_dt_xs <- periodogram(activity, dams_sample, FUN=chi_sq_periodogram)
per_dt_xs_with_peaks <- find_peaks(per_dt_xs, alpha=1e-3)
per_dt_xs_with_peaks[peak==1]
\dontrun{
ggetho::ggperio(per_dt_xs_with_peaks) + geom_line()  +
geom_line(aes(y=signif_threshold), colour="blue") +
geom_point(data = per_dt_xs_with_peaks[peak==1], col="red") +
facet_wrap( ~ id, ncol = 8, labeller = id_labeller)
}
}
