% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_periodogram.R
\name{fourier_periodogram}
\alias{fourier_periodogram}
\title{Fourier periodogram}
\usage{
fourier_periodogram(x, period_range = c(hours(16), hours(32)),
  sampling_rate = 1/mins(1), alpha = 0.05)
}
\arguments{
\item{x}{numeric vector}

\item{period_range}{vector of size 2 defining minimal and maximal range of period to study (in seconds)}

\item{sampling_rate}{the -- implicitly regular -- sampling rate of x (in hertz)}

\item{alpha}{significance level}
}
\value{
\link{data.table} ... todo
}
\description{
TODO
}
\examples{
x <- behavr::toy_dam_data(duration = days(20))$activity
s <- fourier_periodogram(x,  period_range = c(hours(08), hours(18)))
s[, period := period/hours(1)]
# to plot:
# ggplot2::ggplot(s, ggplot2::aes(period, power)) + ggplot2::geom_line()
s <- fourier_periodogram(x, sampling_rate= 1/mins(2),  period_range = c(hours(08) * 2, hours(18) * 2))
s[, period := period/hours(1)]
# to plot:
# ggplot2::ggplot(s, ggplot2::aes(period, power)) + ggplot2::geom_line()
}
