% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls-periodogram.R
\name{ls_periodogram}
\alias{ls_periodogram}
\title{Lomb-Scargle Periodogram}
\usage{
ls_periodogram(x, period_range = c(hours(16), hours(32)),
  sampling_rate = 1/mins(1), alpha = 0.05, oversampling = 8)
}
\arguments{
\item{x}{numeric vector}

\item{period_range}{vector of size 2 defining minimal and maximal range of period to study (in seconds)}

\item{sampling_rate}{the -- implicitly regular -- sampling rate of x (in hertz)}

\item{alpha}{significance level}

\item{oversampling}{the oversampling factor}
}
\description{
TODO
}
\examples{
x <- behavr::toy_dam_data(duration = days(5))$activity
s <- ls_periodogram(x,  period_range = c(hours(08), hours(18)))
# ggplot2::ggplot(s, ggplot2::aes(period, power)) + ggplot2::geom_line() + ggplot2::scale_x_time()
}
\seealso{
\link[lomb:lsp]{lomb::lsp} the orginal function
}
