% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Computes periodograms}
\usage{
periodogram(var, data, period_range = c(hours(16), hours(32)),
  resample_rate = 1/mins(1), alpha = 0.05, FUN = chi_sq_periodogram, ...)
}
\arguments{
\item{var}{variable to analyse}

\item{data}{\link{behavr} table}

\item{period_range}{vector of size 2 defining minimal and maximal range of period to study (in seconds)}

\item{resample_rate}{frequency to resample (up or down) the data at (in hertz)}

\item{alpha}{significance level}

\item{FUN}{function used to compute periodogram (see \link{periodogram_methods})}

\item{...}{additional arguments to be passed to FUN}
}
\value{
a \link{behavr} table with TODO
}
\description{
This function builds peroidograms, with one of several methods, for each individual of \link{behavr} table
}
\examples{
data(dams_sample)
pdt <- periodogram(activity, dams_sample, FUN=ls_periodogram, oversampling = 4)
pdt <- periodogram(activity, dams_sample, FUN=chi_sq_periodogram)
}
